fix: Resolve RuboCop linting issues for ARI WebSocket client (#53)
- Refactor WebSocket class to address Metrics/ClassLength violation
  - Extract connection management into Connection module
  - Extract event handling into EventHandlers module
  - Reduce main class from 268 to 125 lines
- Dependencies already sorted alphabetically (no changes needed)
